+++ src/components/AddTodo/Modal/Modal.module.scss
@@ -44,7 +44,7 @@
   overflow: hidden;
   background-color: #6e6e6e;
   border-radius: 100%;
-  transition: all 1s ease-out;
+  transition: all 0.5s ease-out;
 
   &.active {
     right: 0;
