+++ src/components/Animation/index.js
@@ -6,7 +6,7 @@ import styles from './Animation.module.scss'
 import Menu from '../Menu'
 import MainPage from '../MainPage'
 
-function Animation({ handleModal, tasks, setTasks }) {
+function Animation() {
   const [isOpen, setIsOpen] = useState(false)
   const [userName, setUserName] = useState('Joy')
 
@@ -32,9 +32,6 @@ function Animation({ handleModal, tasks, setTasks }) {
             handleMenuBtnClick={handleMenuBtn}
             userName={userName}
             handleUserNameChange={handleUserNameChange}
-            handleModal={handleModal}
-            tasks={tasks}
-            setTasks={setTasks}
           />
         </div>
       </motion.div>
